First commit project structure license specs and some code
First commit of the project. It has a project structure.

specs file is added to keep in mind the structure that
the projects would be using to organize assets.

Idchecker which checks for duplicate for the ids
of the related documents.

Projectmaker helps to build up a directory structure
that the suite can assume and use for other scripts

validator has only a function to check simple authority
document but it will be extended to cover all the document types
in specs.