New tests are added.
New tests are added and test assets are created.
Tests for validating structure of files are now done.
Recommendations are added to spec file with rfc notice
on the wording.
Small change was introduced to validator to ensure
that string keys like "mykey" can be differentiated from
"0". It is expected that string digits are reserved
for array indices, and not for any ids in any of the
documents produced by the suite.

TODO:

- Follow recommendations.

It is better to introduce a core relations authority file,
then maybe a combined one. Each simple authority file
defining a set of relations should be specific to a
domain codomain pair, so that the relations are
differentiable mathematically. This includes
qualifiers for representations of phenomena.