Spec, container and model change modifications on primitives
Equality, inequality and hash methods are added to
all of the objects in container and in primitive
as well as in model.
Tests of container and model cover basic functionality
but not edge cases.
They should be increased to cover the edge cases as well.
Especially the newly added equality methods and
hash methods needs to be tested, truely.
I might modify so that inequality occurs when
types of object do not conform to each other.
But that might not work so well with hashes.
Some objects might end up with same hash but with
different types etc in the case of inheritence,
or not. I simply need to test this.

Some containers are not needed. They need to be
removed from spec and container.py. But they
show well how the hierarchy of things are constructed
so I might left them as well.
I will also rename model.py to structure.py.
model.py should contain the overall document models like
relation, simple authority, combined authority, predicate
etc.
structure.py should have a corresponding renderer or
viewer.py which transforms structures to different
representations like json/html/xml and the like.
Then there should be a view.py which builds
the models using the rendering functions in
renderer.

Controller should deal with io and creating new
structures or models from old structures.
Since everything is immutable, it is important
to deal with copying and data sharing in an
efficient manner.