Rendering logic change to io logic
Decided to change the application logic once more
to accomodate better io before rendering.

Here is the catch. Rendering should be flexible
enough for accomodating different needs,
however serialization needs to conserve as much
data as possible.

However project data should be exchangeable
using common formats as well. Once the data
underlying the project is obtained, with the
help of several functions, the data can
be used for rendering more popular and easier
to read formats based on needs. Like
skipping certain entity types or predicate
types, etc. Or making it available in html
format, or in a given template etc.

The serialization should be done using
xml, json and yaml. The idea is that
from these files we can build up all the data
and its relation to each other. The serialization
format shows not only value of each primitive or
container in question but also shows the class/
type of the value with it.

Some classes like Constraint String, or other
classes that use functions for evaluating
state of data store function names a seperate
functions file can be shipped with project data
to check the validity of the structure of
the data in the case of a modification of
the project data. Or the functions itself
can be serialized into data of the project
to be evaluated upon reading of the serialized
data.

Io classes need to be updated to conform the
new structure of PrimitiveIo