Changed travis yaml for new tests primitive io done container io started
Travis yaml is changed to accomodate the new tests.
For personal shortcut there is now a test runner py
which simply runs all tests

test_io is removed, now each part of io is going to
be tested individually.
Primitive io is done. I can now do input and output
from and to json and xml

Container is io is work in progress, but the
structure is more or less the same for most of
the containers. However the exact details of
representations have to be thought out for both of
them whic takes a little time.

As per stated with previous commit the rendering
and serializing are treated in different fashion.

Test for container io takes time but still once it is
done it would be a good and reliable tool for production
use.

Once the container io is done, I should concentrate on
structure io. How a certain structure is going to be
serialized is completely independent of how it is going to
be modified etc. As long as it is valid, we serialize it.