Started reimplementing containers to accomodate immutable objects
In order to ensure immutability of containers and primitives
All of them subclasses named tuple.

Primitive tests pass.
Primitive io tests pass as well.

Container tests fail
Container io tests fail as well.

Structures should be adapted to by subclassing them from
named tuple as well.

A maker object is added to both primitive and container.
However container maker is not done yet. Once it is done,
tests should be adapted as well as the io part, which
is not yet complete anyway.