Assertions changed to if nots
Assertions are changed to if not expressions to
improve error handling.
Now we throw a type error, or value error for
related cases.

All primitives are now ready to be used I think.
Containers need some handling in their tests do not
pass yet. Their assert statements are transformed
into if not expressions as well, so their tests
need to be changed as well, this should not make
a difference for their io though.