Radical Changes in spec
This commit introduces radical changes to spec. Most of the containers are now removed. 
We use only Pair and Array as containers now. The idea is to enforce data correctness
using subclasses instead of defining all possible combinations as a class.
Pair captures a Tuple with two elements of different type.
Array captures an array with all elements having the same type.